illumination intensity test

+++ lightcontroller.py
@@ -1,4 +1,7 @@
 from time import sleep
+
+import numpy as np
+
 try:
     from apa102_pi.driver import apa102
     import threading
@@ -24,7 +27,7 @@ try:
         #channel num is 1 to 3 not 0 to 2
         def set_led(self, maze_num, channel_num, red, green, blue, bright_pct = 100):
             with self.lock:
-                self.strip.set_pixel(3*(maze_num-1)+channel_num-1,red, green, blue, bright_percent=bright_pct)
+                self.strip.set_pixel(3*(maze_num-1)+channel_num-1,np.uint8(red), np.uint8(green), np.uint8(blue), bright_percent=bright_pct)
                 print(f"{maze_num}, {channel_num}: {red},{green},{blue}")
 
         def update_leds(self):
